fix(v2.0.18): SQLSTATE 42725 — peak_pnl GREATEST used untyped param arithmetic, silently breaking EVERY paper mark since v2.0.13
Root cause of the 2026-08-20 supervision incident, caught by the v2.0.17
telemetry within 90 seconds of deploy:

  peak_pnl_usdt = GREATEST(peak_pnl_usdt, $3 + $4)
  -> ERROR: operator is not unique: unknown + unknown (SQLSTATE 42725)

PostgreSQL cannot resolve '+' over two untyped (unknown) parameters. The
mark UPDATE failed on EVERY bot on EVERY manage pass since v2.0.13 shipped
the peak column (~04:40Z) — silently, because the Exec was `_, _ =`. What
still worked: closes/adjusts/funding anchors (no param arithmetic), PEPE's
SL (in-memory math per pass), the KAITO tranche-2 (in-memory). What was
lost: all pair-profit booking, mark-to-market, persisted peaks (trailing),
peak-based TP checks — for 6 hours, invisible.

- both GREATEST clauses (paper mark + REAL reconcile) now cast
  $n::NUMERIC + $m::NUMERIC
- swept the entire backend for other $n + $m param arithmetic: none remain

+++ backend/internal/autogrid/worker.go
@@ -2026,7 +2026,7 @@ func (worker *Worker) reconcileAndManage(ctx context.Context) (int, error) {
 				END,
 			    last_remote_status = $4, realized_pnl_usdt = $5,
 			    unrealized_pnl_usdt = $6,
-			    peak_pnl_usdt = GREATEST(COALESCE(peak_pnl_usdt, 0), $5 + $6),
+			    peak_pnl_usdt = GREATEST(COALESCE(peak_pnl_usdt, 0), $5::NUMERIC + $6::NUMERIC),
 			    last_reconciled_at = NOW(),
 			    last_error = NULL, updated_at = NOW()
 			WHERE id = $1
@@ -2814,7 +2814,7 @@ func (worker *Worker) managePaperBots(ctx context.Context, settings Settings) er
 			UPDATE paper_grid_bots
 			SET mark_price = $2, unrealized_pnl_usdt = $3,
 			    realized_pnl_usdt = $4, last_grid_level = $5,
-			    peak_pnl_usdt = GREATEST(peak_pnl_usdt, $3 + $4),
+			    peak_pnl_usdt = GREATEST(peak_pnl_usdt, $3::NUMERIC + $4::NUMERIC),
 			    updated_at = NOW()
 			WHERE id = $1
 		`, bot.id, price, unrealized, realized, currentLevel); mErr != nil {
